requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
redis==5.0.1
Werkzeug==2.3.7